+++ src/prelude.rs
@@ -9,7 +9,7 @@ pub use super::{
 pub use crate::pico8::{
     btn, btnp, camera, canvas::cls, canvas::pset, circ, circfill, color, fget, fillp, fset, line,
     map, mget, mset, music, oval, ovalfill, pal, palm, paln, palt, print::cursor, print::print,
-    rect, rectfill, sfx, sget, spr, sset, sspr,
+    rect, rectfill, rrect, rrectfill, sfx, sget, spr, sset, sspr,
 };
 
 pub use std::str::FromStr;